+++ decidim-core/spec/system/user_activity_spec.rb
@@ -40,6 +40,14 @@ describe "User activity", type: :system do
     create(:dummy_resource, component: component, published_at: Time.current)
   end
 
+  let!(:resource_types) do
+    Decidim::ActionLog
+      .select(:resource_type).distinct
+      .pluck(:resource_type)
+      .map { |r| r.split("::").last }
+      .reject { |r| r.match?(/ParticipatoryProcess|Component|Survey|Result|Assembly|Consultation|DummyResource|missing\stranslation/i) }
+  end
+
   before do
     switch_to_host organization.host
   end
@@ -59,11 +67,11 @@ describe "User activity", type: :system do
       end
     end
 
-    it "displays activities filter" do
+    it "displays activities filter with the correct options" do
       expect(page).to have_select(
         "filter[resource_type]",
         selected: "All types",
-        with_options: ["All types"]
+        options: resource_types.push("All types")
       )
     end
   end